nutellove/requirements.txt
colorama==0.3.9
dj-database-url==0.5.0
Django==2.0.3
django-bootstrap-pagination==1.6.4
django-debug-toolbar==1.9.1
django-widget-tweaks==1.4.1
gunicorn==19.7.1
numpy==1.14.2
pandas==0.22.0
psycopg2==2.7.4
python-dateutil==2.7.0
pytz==2018.3
requests==2.11.1
six==1.11.0
sqlparse==0.2.4
termcolor==1.1.0
tqdm==4.19.7
whitenoise==3.3.1
